@@ -95,91 +21,162 @@
         "spec:integration": "cypress open",
         "clean": "rimraf build"
     },
+    "dependencies": {
+        "@capacitor/android": "2.4.1",
+        "@capacitor/cli": "^2.4.2",
+        "@capacitor/core": "^2.4.1",
+        "@capacitor/ios": "2.4.1",
+        "@ethersproject/providers": "^5.7.2",
+        "@formatjs/intl-pluralrules": "^5.1.8",
+        "@formatjs/intl-relativetimeformat": "^11.1.8",
+        "@ionic/cli": "^6.11.11",
+        "@openware/cryptofont": "2.3.0",
+        "@openware/react-components": "^0.0.5",
+        "@openware/react-opendax": "0.1.37",
+        "@sentry/browser": "^7.37.1",
+        "@types/ws": "^8.5.4",
+        "@typescript-eslint/eslint-plugin": "^5.51.0",
+        "@typescript-eslint/parser": "^5.51.0",
+        "@web3-react/core": "^6.1.9",
+        "@web3-react/injected-connector": "^6.0.7",
+        "axios": "^1.3.2",
+        "bitcoincash-regex": "^1.1.9",
+        "bootstrap": "^4.3.1",
+        "css-loader": "6.7.3",
+        "eslint-config-prettier": "^8.6.0",
+        "eslint-plugin-prettier": "^4.2.1",
+        "eslint-plugin-react": "^7.32.2",
+        "flat": "^5.0.2",
+        "history": "^4.7.2",
+        "http-signature": "^1.3.6",
+        "i18n-iso-countries": "^7.5.0",
+        "immutability-helper": "^3.1.1",
+        "js-yaml": "^4.1.0",
+        "moment": "^2.29.4",
+        "moment-timezone": "^0.5.40",
+        "multicoin-address-validator": "^0.5.12",
+        "node-sass": "^8.0.0",
+        "postcss-cli": "^10.1.0",
+        "prettier": "^2.8.4",
+        "qrcode.react": "3.1.0",
+        "react": "^18.2.0",
+        "react-bootstrap": "^1.0.0-beta.16",
+        "react-color": "^2.19.3",
+        "react-dom": "^18.2.0",
+        "react-fade-in": "^2.0.1",
+        "react-ga": "^3.3.1",
+        "react-google-recaptcha": "^1.0.4",
+        "react-grid-layout": "^1.3.4",
+        "react-intl": "^6.2.8",
+        "react-maskinput": "^3.2.0",
+        "react-perfect-scrollbar": "^1.5.8",
+        "react-redux": "^8.0.5",
+        "react-router": "^5.1.2",
+        "react-router-dom": "^5.1.2",
+        "react-scroll": "^1.8.9",
+        "react-select": "^5.7.0",
+        "react-use-websocket": "^4.3.1",
+        "recharts": "^2.4.1",
+        "redux": "^4.2.1",
+        "redux-saga": "^1.2.2",
+        "redux-sentry-middleware": "^0.2.2",
+        "sinon": "^7.2.3",
+        "style-loader": "^3.3.1",
+        "stylelint": "^15.1.0",
+        "stylelint-config-standard": "^30.0.0",
+        "stylelint-no-unsupported-browser-features": "^6.0.0",
+        "stylelint-order": "^6.0.0",
+        "stylelint-prettier": "^2.0.0",
+        "stylelint-scss": "^4.4.0",
+        "stylelint-selector-bem-pattern": "^2.1.0",
+        "thread-loader": "^3.0.4",
+        "ts-loader": "^9.4.2",
+        "ua-parser-js": "^1.0.33",
+        "url-loader": "^4.1.1",
+        "ws": "^8.12.0"
+    },
     "devDependencies": {
+        "@babel/core": "^7.20.12",
         "@openware/coding-standards": "2.3.0",
-        "@types/axios": "^0.14.0",
-        "@types/classnames": "^2.2.6",
-        "@types/copy-webpack-plugin": "^6.0.0",
         "@types/enzyme": "^3.1.13",
         "@types/enzyme-adapter-react-16": "^1.0.3",
         "@types/history": "^4.7.0",
         "@types/jest": "^26.0.10",
         "@types/jquery": "^3.3.6",
-        "@types/mini-css-extract-plugin": "^0.9.1",
+        "@types/mini-css-extract-plugin": "^2.5.1",
         "@types/node": "^10.9.4",
         "@types/qrcode.react": "^1.0.1",
-        "@types/react": "^16.9.14",
-        "@types/react-dom": "^16.9.4",
-        "@types/react-hot-loader": "^4.1.0",
+        "@types/react": "^18.0.27",
+        "@types/react-dom": "^18.0.10",
         "@types/react-redux": "^7.1.9",
         "@types/react-router": "^5.1.11",
         "@types/react-router-dom": "^5.1.3",
         "@types/react-test-renderer": "^16.0.2",
         "@types/redux-form": "^8.2.7",
-        "@types/webpack": "^4.41.21",
-        "@types/webpack-dev-server": "^3.11.0",
-        "@types/webpack-env": "^1.15.2",
-        "@types/webpack-merge": "^4.1.5",
+        "@types/webpack": "^5.28.0",
+        "@types/webpack-env": "^1.18.0",
         "autoprefixer": "^10.3.1",
         "axios-mock-adapter": "^1.16.0",
-        "chalk": "^2.4.1",
-        "classnames": "^2.2.6",
-        "colors": "1.3.2",
-        "copy-webpack-plugin": "^6.2.1",
+        "chalk": "^5.2.0",
+        "classnames": "^2.3.2",
+        "colors": "1.4.0",
+        "copy-webpack-plugin": "^11.0.0",
         "cross-env": "^7.0.2",
         "cryptocurrency-icons": "^0.16.1",
-        "cssnano": "^4.0.5",
+        "css-minimizer-webpack-plugin": "^4.2.2",
+        "cssnano": "^5.1.14",
         "cypress": "^3.1.5",
-        "dompurify": "^2.3.1",
+        "dompurify": "^3.0.0",
         "enzyme": "^3.10.0",
         "enzyme-adapter-react-16": "^1.10.0",
         "enzyme-to-json": "^3.3.4",
-        "eslint": "^7.7.0",
-        "eslint-import-resolver-typescript": "^2.2.1",
-        "eslint-import-resolver-webpack": "^0.12.2",
-        "eslint-plugin-import": "^2.22.0",
-        "file-loader": "^1.1.11",
-        "fork-ts-checker-webpack-plugin": "^5.1.0",
+        "eslint": "^8.34.0",
+        "eslint-import-resolver-typescript": "^3.5.3",
+        "eslint-import-resolver-webpack": "^0.13.2",
+        "eslint-plugin-import": "^2.27.5",
+        "file-loader": "6.2.0",
+        "fork-ts-checker-webpack-plugin": "^7.3.0",
         "header-case-normalizer": "1.0.3",
-        "html-webpack-plugin": "^4.3.0",
-        "husky": "^1.0.0-rc.13",
-        "isomorphic-ws": "^4.0.1",
+        "html-webpack-plugin": "^5.5.0",
+        "husky": "^8.0.3",
+        "isomorphic-ws": "^5.0.0",
+        "javascript-obfuscator": "^4.0.2",
         "jest": "^26.4.1",
         "jest-config": "^26.4.1",
         "jest-environment-enzyme": "^7.1.2",
         "jest-enzyme": "^7.1.2",
-        "js-combinatorics": "0.5.0",
+        "js-combinatorics": "2.1.1",
         "lodash-webpack-plugin": "^0.11.6",
-        "mini-css-extract-plugin": "^0.10.0",
+        "mini-css-extract-plugin": "^2.7.2",
         "npm-run-all": "^4.1.5",
-        "optimize-css-assets-webpack-plugin": "^5.0.3",
-        "postcss": "^8.1.4",
-        "postcss-autoreset": "^2.0.2",
-        "postcss-custom-properties": "^10.0.0",
-        "postcss-hexrgba": "^2.0.1",
-        "postcss-import": "^13.0.0",
-        "postcss-loader": "^4.0.4",
-        "postcss-nested": "^5.0.1",
-        "ramda": "^0.26.1",
-        "react-hot-loader": "^4.3.4",
-        "react-test-renderer": "^16.4.2",
-        "redux-mock-store": "^1.5.3",
-        "rimraf": "^2.6.2",
-        "sass-loader": "10.2.0",
-        "source-map-explorer": "^1.7.0",
-        "speed-measure-webpack-plugin": "^1.3.3",
-        "stylelint-config-prettier": "^8.0.2",
-        "stylelint-config-recommended": "^3.0.0",
+        "postcss": "^8.4.21",
+        "postcss-autoreset": "^3.0.4",
+        "postcss-custom-properties": "^13.1.3",
+        "postcss-hexrgba": "^2.1.0",
+        "postcss-import": "^15.0.0",
+        "postcss-loader": "^7.0.2",
+        "postcss-nested": "^6.0.0",
+        "prettier-plugin-organize-imports": "^3.2.2",
+        "prop-types": "^15.8.1",
+        "ramda": "^0.28.0",
+        "react-test-renderer": "^18.2.0",
+        "redux-mock-store": "^1.5.4",
+        "rimraf": "^4.1.2",
+        "sass-loader": "13.2.0",
+        "source-map-explorer": "^2.5.3",
+        "speed-measure-webpack-plugin": "^1.5.0",
+        "stylelint-config-prettier": "^9.0.5",
+        "stylelint-config-recommended": "^10.0.0",
         "tailwindcss": "^2.2.7",
-        "ts-jest": "^26.2.0",
-        "ts-node": "^9.0.0",
-        "typescript": "^3.9.7",
-        "webpack": "^4.44.1",
-        "webpack-bundle-analyzer": "^4.4.0",
-        "webpack-cli": "^3.3.12",
-        "webpack-dev-server": "^3.11.0",
-        "webpack-merge": "^5.1.2",
-        "webpack-obfuscator": "^0.18.0"
+        "ts-jest": "^29.0.5",
+        "ts-node": "^10.9.1",
+        "typescript": "^4.9.5",
+        "webpack": "^5.75.0",
+        "webpack-bundle-analyzer": "^4.8.0",
+        "webpack-cli": "^5.0.1",
+        "webpack-dev-server": "^4.11.1",
+        "webpack-merge": "^5.8.0",
+        "webpack-obfuscator": "^3.5.1"
     },
     "browserslist": [
         ">0.2%",
