@@ -71,7 +71,7 @@ class ProfileAuthDetailsComponent extends React.Component<Props, State> {
         };
     }
 
-    public componentWillReceiveProps(next: Props) {
+    public UNSAFE_componentWillReceiveProps(next: Props) {
         if (next.passwordChangeSuccess) {
             this.setState({ showChangeModal: false });
             this.props.changePasswordReset();
