@@ -69,7 +69,7 @@ class ChangeForgottenPasswordComponent extends React.Component<Props, ChangeForg
         }
     }
 
-    public componentWillReceiveProps(next: Props) {
+    public UNSAFE_componentWillReceiveProps(next: Props) {
         if (next.changeForgotPassword && (!this.props.changeForgotPassword)) {
             this.props.history.push('/signin');
         }
