@@ -24,7 +24,7 @@ class Maintenance extends React.Component<Props> {
         }
     }
 
-    public componentWillReceiveProps(nextProps: Props) {
+    public UNSAFE_componentWillReceiveProps(nextProps: Props) {
         if (!this.props.status.length && nextProps.status.length && nextProps.status !== 'maintenance') {
             this.props.history.replace('/');
         }
