@@ -1,7 +1,5 @@
 const tailwindcss = require('tailwindcss');
 
 module.exports = {
-  plugins: [
-    tailwindcss,
-  ]
-}
+    plugins: [tailwindcss, require('autoprefixer'), require('postcss-nested')],
+};
