@@ -1,5 +1,5 @@
-import * as React from 'react';
-import Select, { ValueType } from 'react-select';
+import * as React from "react";
+import Select, { OnChangeValue } from "react-select";
 
 export interface OptionsInterface {
     label: string;
@@ -8,7 +8,7 @@ export interface OptionsInterface {
 
 interface OwnProps {
     options: OptionsInterface[];
-    onSelect: (option: ValueType<OptionsInterface, false>) => void;
+    onSelect: (option: OnChangeValue<OptionsInterface, false>) => void;
     placeholder: string;
     className?: string;
     classNamePrefix?: string;
@@ -17,7 +17,7 @@ interface OwnProps {
 export const SearchDropdown: React.FC<OwnProps> = ({ className, onSelect, options, placeholder, classNamePrefix }) => (
     <Select
         className={className}
-        classNamePrefix={classNamePrefix || 'cr-search-dropdown'}
+        classNamePrefix={classNamePrefix || "cr-search-dropdown"}
         options={options}
         onChange={onSelect}
         placeholder={placeholder}
