@@ -1,20 +1,20 @@
-import Axios from 'axios';
-import MockAdapter from 'axios-mock-adapter';
-import { Action, Middleware } from 'redux';
-import configureMockStore from 'redux-mock-store';
-import { Cryptobase } from '../api';
+import Axios from "axios";
+import MockAdapter from "axios-mock-adapter";
+import { Action, Middleware } from "redux";
+import configureMockStore from "redux-mock-store";
+import { Cryptobase } from "../api";
 
 // tslint:disable-next-line
-import * as WebSocket from 'ws';
+import * as WebSocket from "ws";
 
 const mockConfig: Config = {
     api: {
-        authUrl: '/api/v2/barong',
-        tradeUrl: '/api/v2/peatio',
-        applogicUrl: '/api/v2/applogic',
-        rangerUrl: '/api/v2/ranger',
-        finexUrl: '/api/v2/finex',
-        p2pUrl: '/api/v2/p2p',
+        authUrl: "/api/v2/barong",
+        tradeUrl: "/api/v2/peatio",
+        applogicUrl: "/api/v2/applogic",
+        rangerUrl: "/api/v2/ranger",
+        finexUrl: "/api/v2/finex",
+        p2pUrl: "/api/v2/p2p",
     },
     finex: false,
     withCredentials: false,
@@ -25,33 +25,28 @@ const mockConfig: Config = {
     sentryEnabled: false,
     captchaLogin: false,
     usernameEnabled: false,
-    gaTrackerKey: '',
-    minutesUntilAutoLogout: '5',
-    msAlertDisplayTime: '5000',
-    msPricesUpdates: '1000',
-    sessionCheckInterval: '15000',
-    balancesFetchInterval: '3000',
-    passwordEntropyStep: '14',
-    password_min_entropy: '',
-    captcha_type: 'none',
+    gaTrackerKey: "",
+    minutesUntilAutoLogout: "5",
+    msAlertDisplayTime: "5000",
+    msPricesUpdates: "1000",
+    sessionCheckInterval: "15000",
+    balancesFetchInterval: "3000",
+    passwordEntropyStep: "14",
+    password_min_entropy: "",
+    captcha_type: "none",
     storage: {
-        defaultStorageLimit: '50',
-        orderBookSideLimit: '25'
+        defaultStorageLimit: "50",
+        orderBookSideLimit: "25",
     },
-    languages: ['en', 'ru'],
-    kycSteps: [
-        'email',
-        'phone',
-        'profile',
-        'document',
-        'address'
-    ],
-    themeSwitcher: 'visible',
-    useSharedLayout: 'false',
+    languages: ["en", "ru"],
+    kycSteps: ["email", "phone", "profile", "document", "address"],
+    themeSwitcher: "visible",
+    useSharedLayout: "false",
+    tvDefaultCandles: 20,
 };
 
 // tslint:disable no-any no-console
-export const loggerMiddleware: Middleware = (store: {}) => (next: any) => (action: Action) => {
+export const loggerMiddleware: Middleware = (store: object) => (next: any) => (action: Action) => {
     console.log(`dispatching: ${JSON.stringify(action)}`);
 
     return next(action);
@@ -75,13 +70,13 @@ export const mockNetworkError = (mockAxios: any) => {
 
 export const createEchoServer = (port: number, debug: boolean) => {
     const server = new WebSocket.Server({ port: port });
-    server.on('connection', (ws, request) => {
+    server.on("connection", (ws, request) => {
         if (debug) {
-            ws.addEventListener('open', () => {
+            ws.addEventListener("open", () => {
                 console.log(`Ping Server: listening on port ${port}`);
             });
         }
-        ws.on('message', (message: string) => {
+        ws.on("message", (message: string) => {
             if (debug) {
                 console.log(`Ping Server: sending back ${message}`);
             }
