@@ -1,4 +1,3 @@
 export const AppConfig = {
-    api: {
-    }
+    api: {},
 };
