@@ -1,16 +1,30 @@
-import { HotModuleReplacementPlugin } from 'webpack';
-import MiniCssExtractPlugin from 'mini-css-extract-plugin';
-import HtmlWebpackPlugin from 'html-webpack-plugin';
-import path from 'path';
-import merge from 'webpack-merge';
-import 'webpack-dev-server';
 import ForkTsCheckerWebpackPlugin from 'fork-ts-checker-webpack-plugin';
-
+import HtmlWebpackPlugin from 'html-webpack-plugin';
+// eslint-disable-next-line import/order
+import path from 'path';
+import { Configuration, HotModuleReplacementPlugin } from 'webpack';
+import 'webpack-dev-server';
+import type { Configuration as DevServerConfiguration } from 'webpack-dev-server';
+// eslint-disable-next-line import/no-named-as-default
+import merge from 'webpack-merge';
 import commonConfig from './common';
 const rootDir = path.resolve(__dirname, '..');
 
-const config = merge(commonConfig, {
-    devtool: 'cheap-module-eval-source-map',
+const devServer: DevServerConfiguration = {
+    compress: false,
+    port: 3000,
+    historyApiFallback: true,
+    static: {
+        directory: path.join(__dirname, '../public'),
+    },
+};
+
+const config: Configuration = merge(commonConfig, {
+    mode: 'development',
+    output: {
+        pathinfo: false,
+    },
+    devtool: 'eval-cheap-module-source-map',
     plugins: [
         new HotModuleReplacementPlugin(),
         new ForkTsCheckerWebpackPlugin({}),
@@ -20,6 +34,12 @@ const config = merge(commonConfig, {
             chunks: ['common', 'bundle', 'styles'],
         }),
     ],
+    optimization: {
+        runtimeChunk: true,
+        removeAvailableModules: false,
+        removeEmptyChunks: false,
+        splitChunks: false,
+    },
     module: {
         rules: [
             {
@@ -28,25 +48,11 @@ const config = merge(commonConfig, {
             },
             {
                 test: /\.(css|sass|scss|pcss)$/,
-                use: [
-                    'style-loader',
-                    {
-                        loader: MiniCssExtractPlugin.loader,
-                        options: {
-                            hmr: true,
-                            reloadAll: true,
-                        },
-                    },
-                    'cache-loader',
-                    'css-loader',
-                    'sass-loader',
-                    'postcss-loader',
-                ],
+                use: ['style-loader', 'css-loader', 'sass-loader', 'postcss-loader'],
             },
             {
                 test: /\.(tsx|ts)?$/,
                 use: [
-                    'cache-loader',
                     {
                         loader: 'thread-loader',
                         options: {
@@ -60,23 +66,12 @@ const config = merge(commonConfig, {
                             happyPackMode: true,
                         },
                     },
-                    
                 ],
-                
                 exclude: /node_modules/,
             },
         ],
     },
-    devServer: {
-        contentBase: path.join(__dirname, '../public'),
-        compress: false,
-        port: 3000,
-        historyApiFallback: true,
-        stats: {
-            children: false,
-        },
-        hot: true,
-    },
+    devServer,
 });
 
 // eslint-disable-next-line import/no-default-export
