@@ -1,34 +1,37 @@
-import { ExtendedAPIPlugin, DefinePlugin } from 'webpack';
-import MiniCssExtractPlugin from 'mini-css-extract-plugin';
-import merge from 'webpack-merge';
-import OptimizeCssAssetsPlugin from 'optimize-css-assets-webpack-plugin';
-import path from 'path';
 import CopyWebpackPlugin from 'copy-webpack-plugin';
-import JavaScriptObfuscator from 'webpack-obfuscator';
+import CssMinimizerPlugin from 'css-minimizer-webpack-plugin';
+import HtmlWebpackPlugin from 'html-webpack-plugin';
+import MiniCssExtractPlugin from 'mini-css-extract-plugin';
+// eslint-disable-next-line import/order
+import path from 'path';
+import { DefinePlugin } from 'webpack';
+// eslint-disable-next-line import/no-named-as-default
+import merge from 'webpack-merge';
+import WebpackObfuscator from 'webpack-obfuscator';
+import commonConfig from './common';
+// eslint-disable-next-line @typescript-eslint/no-var-requires
 const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
 
 const rootDir = path.resolve(__dirname, '..');
 const BUILD_DIR = path.resolve(rootDir, 'build');
 
-import commonConfig from './common';
-
 const domain = process.env.BUILD_DOMAIN ? process.env.BUILD_DOMAIN.split(',') : [];
 
 const plugins = [
-    new ExtendedAPIPlugin(),
     new DefinePlugin({ 'process.env.BUILD_EXPIRE': JSON.stringify(process.env.BUILD_EXPIRE) }),
-    new OptimizeCssAssetsPlugin({
-        assetNameRegExp: /\.css$/g,
-        cssProcessor: require('cssnano'),
-        cssProcessorPluginOptions: {
-            preset: ['default', { discardComments: { removeAll: true } }],
-        },
-        canPrint: false,
+    new HtmlWebpackPlugin({
+        template: path.resolve(rootDir, 'src/app/template.html'),
+        hash: true,
+        chunks: ['common', 'bundle', 'styles'],
+    }),
+    new MiniCssExtractPlugin({
+        filename: '[name].[contenthash].css',
+        chunkFilename: '[id].[contenthash].css',
     }),
     new CopyWebpackPlugin({
         patterns: [{ from: 'public' }],
     }),
-    new JavaScriptObfuscator({ rotateUnicodeArray: true, domainLock: domain }),
+    new WebpackObfuscator({ rotateStringArray: true, domainLock: domain }),
 ];
 
 if (process.env.ANALYZE === '1') {
@@ -39,20 +42,19 @@ const config = merge(commonConfig, {
     mode: 'production',
     output: {
         path: BUILD_DIR,
-        filename: '[name].[hash].js',
-        globalObject: 'this',
+        filename: '[name].[contenthash].js',
         publicPath: '/',
     },
     optimization: {
         usedExports: false,
-        minimize: true,
+        minimizer: [new CssMinimizerPlugin()],
     },
     plugins,
     module: {
         rules: [
             {
                 test: /\.(css|sass|scss|pcss)$/,
-                use: ['style-loader', MiniCssExtractPlugin.loader, 'css-loader', 'sass-loader', 'postcss-loader'],
+                use: [MiniCssExtractPlugin.loader, 'css-loader', 'sass-loader', 'postcss-loader'],
             },
             {
                 test: /\.(tsx|ts)?$/,
@@ -75,12 +77,25 @@ const config = merge(commonConfig, {
             },
             {
                 test: /\.(woff|woff2|eot|ttf|otf)$/i,
-                use: [{
-                    loader: 'file-loader',
-                    options: {
-                        outputPath: 'fonts',
+                use: [
+                    {
+                        loader: 'file-loader',
+                        options: {
+                            outputPath: 'fonts',
+                        },
                     },
-                }]
+                ],
+            },
+            {
+                test: /\.(js)?$/,
+                exclude: /node_modules/,
+                enforce: 'post',
+                use: {
+                    loader: WebpackObfuscator.loader,
+                    options: {
+                        rotateStringArray: true,
+                    },
+                },
             },
         ],
     },
